+++ inst/tinytest/testBugs.R
@@ -41,3 +41,63 @@ expect_error_xl(
   "Some of the column names of the assays table are missing",
   info = "It doesn't like the new column name 'id'"
 )
+
+rm(x)
+
+# Custom plots with no assays --------------------------------------------------
+
+# Now that custom plots can include data from the results table, it's no longer
+# required to have assays data.
+
+studyNoAssays <- OmicNavigator:::testStudyMinimal()
+
+plotBetas <- function(x) {
+  plot(x[["results"]][["beta"]])
+}
+
+studyNoAssays <- addPlots(
+  studyNoAssays,
+  plots = list(
+    default = list(
+      plotBetas = list(
+        displayName = "Plot the betas",
+        plotType = "multiFeature"
+      )
+    )
+  )
+)
+
+expect_message_xl(
+  plotStudy(
+    studyNoAssays,
+    modelID = names(studyNoAssays[["results"]])[1],
+    featureID = studyNoAssays[["results"]][[1]][[1]][[1]][1:5], # first 5 featureIDs
+    plotID = "plotBetas",
+    testID = names(studyNoAssays[["results"]][[1]])[1]
+  ),
+  "No assays available for modelID",
+  info = "Send message for missing assays data when testID is defined"
+)
+
+expect_error_xl(
+  plotStudy(
+    studyNoAssays,
+    modelID = names(studyNoAssays[["results"]])[1],
+    featureID = studyNoAssays[["results"]][[1]][[1]][[1]][1:5], # first 5 featureIDs
+    plotID = "plotBetas",
+    testID = NULL
+  ),
+  "Add assays data with addAssays()",
+  info = "Throw error for missing assays data when testID is **not** defined"
+)
+
+expect_true_xl(
+  OmicNavigator:::validatePlots(studyNoAssays)
+)
+
+expect_message_xl(
+  OmicNavigator:::validateStudy(studyNoAssays),
+  "Custom plots often use assays"
+)
+
+rm(studyNoAssays, plotBetas)